feat: refactor reward classifiers to use base utilities and restructure imports
