+++ rl_garden/models/reward/classifiers/base/__init__.py
@@ -1,0 +1,18 @@
+"""Shared utilities for classifier datasets and trainers."""
+
+from rl_garden.models.reward.classifiers.base.io import decode_jpeg, decode_jpeg_crop
+from rl_garden.models.reward.classifiers.base.loop import BaseBinaryClassifierTrainer, BaseTrainConfig
+from rl_garden.models.reward.classifiers.base.metrics import compute_metrics
+from rl_garden.models.reward.classifiers.base.model import ResNetBinaryClassifier
+from rl_garden.models.reward.classifiers.base.transforms import build_transforms, empty_image
+
+__all__ = [
+	"decode_jpeg",
+	"decode_jpeg_crop",
+	"BaseBinaryClassifierTrainer",
+	"BaseTrainConfig",
+	"compute_metrics",
+	"ResNetBinaryClassifier",
+	"build_transforms",
+	"empty_image",
+]
