--- rl_garden/models/reward/classifiers/shared/__init__.py
@@ -1,18 +1,0 @@
-"""Shared utilities for classifier datasets and trainers."""
-
-from rl_garden.models.reward.classifiers.shared.io import decode_jpeg, decode_jpeg_crop
-from rl_garden.models.reward.classifiers.shared.loop import BaseBinaryClassifierTrainer, BaseTrainConfig
-from rl_garden.models.reward.classifiers.shared.metrics import compute_metrics
-from rl_garden.models.reward.classifiers.shared.model import ResNetBinaryClassifier
-from rl_garden.models.reward.classifiers.shared.transforms import build_transforms, empty_image
-
-__all__ = [
-	"decode_jpeg",
-	"decode_jpeg_crop",
-	"BaseBinaryClassifierTrainer",
-	"BaseTrainConfig",
-	"compute_metrics",
-	"ResNetBinaryClassifier",
-	"build_transforms",
-	"empty_image",
-]
